rename reservation script
