+++ show_movies.py
@@ -1,4 +1,4 @@
-from reservation import Reservation
+from magic_reservation_system import Reservation
 
 res = Reservation("cinema.db")
 res.show_movies()
